requirements.txt
# Core Framework
Flask==2.3.3
bcrypt==4.0.1
cryptography==41.0.7
PyJWT==2.8.0

# Machine Learning & Fraud Detection
tensorflow==2.13.0
scikit-learn==1.3.2
numpy==1.24.3
pandas==1.5.3
huggingface-hub==0.17.3

# Communication & Alerts
twilio==8.9.1
gtts==2.4.0

# Database & Storage
sqlite3
json

# Security & Encryption
hashlib
secrets
hmac

# Performance & Optimization
cachetools==5.3.2
redis==5.0.1

# Utilities
requests==2.31.0
python-dotenv==1.0.0
